ReuleauxCRM/requirements.txt
asgiref==3.5.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.11.4
Markdown==3.3.7
oauthlib==3.2.2
Pillow==9.1.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.8.0
